+++ packages/pigeon/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 10.1.2
+
+* [swift] Fixes a crash when passing `null` for nested nullable classes. 
+
 ## 10.1.1
 
 * Updates README to better reflect modern usage.